Create Translation Style Folder and move styles
